requirements.txt
streamlit
feedparser
google-generativeai
openai
newspaper3k
lxml_html_clean